+++ packages/terra-application-links/CHANGELOG.md
@@ -4,6 +4,16 @@ Changelog
 Unreleased
 ----------
 
+6.16.0 - (October 21, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
+6.15.0 - (October 16, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
 6.14.0 - (October 3, 2019)
 ------------------
 ### Changed