+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -3,6 +3,27 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Duplicate ID in examples changed.
+
+4.17.0 - (October 21, 2019)
+------------------
+### Changed
+* Updated WDIO screenshots to account for button updates in terra-date-picker
+
+### Fixed
+* Fixed onBlur event data for entering invalid dates.
+
+4.16.0 - (October 16, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
+### Added
+* Added `isIncomplete` prop and stylings
+* Added `isInvalid` prop and stylings
+* Added `isInvalidMeridiem` prop
+* Added `required` prop
 
 4.15.0 - (October 3, 2019)
 ------------------