+++ packages/terra-dialog-modal/CHANGELOG.md
@@ -3,6 +3,17 @@ Changelog
 
 Unreleased
 ----------
+
+3.17.0 - (October 21, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
+3.16.0 - (October 16, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
 ### Added
 * Added `closeOnOutsideClick` prop to differentiate close on outside click from close by escape key.