+++ packages/terra-slide-panel-manager/CHANGELOG.md
@@ -3,6 +3,19 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Duplicate IDs in examples changed.
+* Jest snapshot updates.
+
+5.15.0 - (October 21, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
+5.14.0 - (October 16, 2019)
+------------------
+### Changed
+* Minor dependency updates
 
 5.13.0 - (October 3, 2019)
 ------------------